♻️ F11: ORDER > Merged BasketProvider into MainProvider
