+++ src/components/pages/order/Main.tsx
@@ -6,18 +6,15 @@ import Basket from "./Basket/Basket";
 import AdminPanel from "./AdminPanel/AdminPanel";
 import { useContext } from "react";
 import { ProductsContext } from "../../../context/OrderMainContext";
-import { BasketProvider } from "./BasketProvider";
 
 export default function Main() {
-  const {menuProds} = useContext(ProductsContext)
+  const { menuProds } = useContext(ProductsContext);
 
   return (
     <MainStyled>
-      <BasketProvider>
-        <Basket products={menuProds} />
-        <Menu />
-        <AdminPanel />
-      </BasketProvider>
+      <Basket products={menuProds} />
+      <Menu />
+      <AdminPanel />
     </MainStyled>
   );
 }
